% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBasicFunctions.R
\name{checkDirectory}
\alias{checkDirectory}
\title{Check if a variable which should contains a path for a directory exists and has a value which corresponds to an existing folder}
\usage{
checkDirectory(variableDirectory, default = NA, isRequired = T)
}
\arguments{
\item{variableDirectory}{a string containing the name of the variable to check}

\item{default}{the result if the variableDirectory is not assigned or the directory does not exists and isRequired is set to FALSE, default is NA}

\item{isRequired}{a logical value indicating if the function should stop in case the folder does not exists, default is TRUE}
}
\value{
The content of \code{variableDirectory} if it is the path to an existing directory or the default value
}
\description{
Check if a variable which should contains a path for a directory exists and has a value which corresponds to an existing folder
}
\details{
If isRequired is set to \code{TRUE} and the content of \code{variableDirectory} is not valid, it will stop and execute an error.
}
\examples{
# A directory which exists:
myDirectory <- "./"
myCheckedDirectory <- checkDirectory("myDirectory")

# A variable which does not exists:
myCheckedDirectory <- checkDirectory("myImaginaryVariableWhichShouldNotExists",isRequired=F)

# A variable which does not exists but with default:
myCheckedDirectory <- checkDirectory("myImaginaryVariableWhichShouldNotExists", default="./", isRequired=F)

# A variable which does not exists but is required:
myCheckedDirectory <- checkDirectory("myImaginaryVariableWhichShouldNotExists")

}
