% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBasicFunctions.R
\name{checkFile}
\alias{checkFile}
\title{Check if a variable which should contains a path for a file exists and has a value which corresponds to an existing file}
\usage{
checkFile(variableFile, default = NA, isRequired = T)
}
\arguments{
\item{variableFile}{a string containing the name of the variable to check}

\item{default}{the result if the variableFile is not assigned or the file does not exists and isRequired is set to FALSE, default is NA}

\item{isRequired}{a logical value indicating if the function should stop in case the file does not exists, default is TRUE}
}
\value{
The content of \code{variableFile} if it is the path to an existing file or the default value
}
\description{
Check if a variable which should contains a path for a file exists and has a value which corresponds to an existing file
}
\details{
If isRequired is set to \code{TRUE} and the content of \code{variableFile} is not valid, it will stop and execute an error.
}
\examples{
# A file which exists:
myFile <- ".Rhistory"
myCheckedFile <- checkFile("myFile")

# A variable which does not exists:
myCheckedFile <- checkFile("myImaginaryVariableWhichShouldNotExists",isRequired=F)

# A variable which does not exists but with default:
myCheckedFile <- checkFile("myImaginaryVariableWhichShouldNotExists", default=".Rhistory", isRequired=F)

# A variable which does not exists but is required:
myCheckedFile <- checkFile("myImaginaryVariableWhichShouldNotExists")

}
