% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBasicFunctions.R
\name{checkLogicalValue}
\alias{checkLogicalValue}
\title{Check if a variable which should contains a logical value exists and its content is logical}
\usage{
checkLogicalValue(variableL, default = NA, isRequired = T)
}
\arguments{
\item{variableL}{a string containing the name of the variable to check}

\item{default}{the result if the \code{variableL} is not assigned or the content are not logical and isRequired is set to FALSE, default is NA}

\item{isRequired}{a logical value indicating if the function should stop in case the values are not numeric, default is TRUE}
}
\value{
The first value contained in \code{variableL} if it is logical or the default value
}
\description{
Check if a variable which should contains a logical value exists and its content is logical
}
\details{
If isRequired is set to \code{TRUE} and the content of \code{variableL} is not valid, it will stop and execute an error.
}
\examples{
# the content is logical:
myLogicalValue <- TRUE
myCheckedLogicalValue <- checkLogicalValue("myLogicalValue")

myLogicalValue <- NA
myCheckedLogicalValue <- checkLogicalValue("myLogicalValue")

# my value is not logical:
myLogicalValue <- "T"
\dontrun{
myCheckedLogicalValue <- checkLogicalValue("myLogicalValue")
}

# A variable which does not exists:
myCheckedLogicalValue <- checkLogicalValue("myImaginaryVariableWhichShouldNotExists",
                                           isRequired = FALSE)

# A variable which does not exists but with default:
myCheckedLogicalValue <- checkLogicalValue("myImaginaryVariableWhichShouldNotExists",
                                           default = TRUE, isRequired = FALSE)

# A variable which does not exists but is required:
\dontrun{
myCheckedLogicalValue <- checkLogicalValue("myImaginaryVariableWhichShouldNotExists")
}
}
