% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBasicFunctions.R
\name{checkNumericalValues}
\alias{checkNumericalValues}
\title{Check if a variable which should contains numerical values exists and all values are numeric}
\usage{
checkNumericalValues(variableN, default = NA, isRequired = T)
}
\arguments{
\item{variableN}{a string containing the name of the variable to check}

\item{default}{the result if the \code{variableN} is not assigned or some values are not numeric and isRequired is set to FALSE, default is NA}

\item{isRequired}{a logical value indicating if the function should stop in case the values are not numeric, default is TRUE}
}
\value{
The content of \code{variableN} if they are all numeric or the default value
}
\description{
Check if a variable which should contains numerical values exists and all values are numeric
}
\details{
If isRequired is set to \code{TRUE} and the content of \code{variableN} is not valid, it will stop and execute an error.
}
\examples{
# all values are numeric:
myNumbers <- 1:3
myCheckedNumbers <- checkNumericalValues("myNumbers")

# some values are not numeric:
myNumbers <- list(1,2,"T")
myCheckedNumbers <- checkNumericalValues("myNumbers")

# A variable which does not exists:
myCheckedNumbers <- checkNumericalValues("myImaginaryVariableWhichShouldNotExists", isRequired = F)

# A variable which does not exists but with default:
myCheckedNumbers <- checkNumericalValues("myImaginaryVariableWhichShouldNotExists", default = c(1,2,12), isRequired = F)

# A variable which does not exists but is required:
myCheckedNumbers <- checkNumericalValues("myImaginaryVariableWhichShouldNotExists")

}
