% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBasicFunctions.R
\name{commonEnd}
\alias{commonEnd}
\title{Find the longest common substring from the end between 2 words}
\usage{
commonEnd(word1, word2)
}
\arguments{
\item{word1}{first string}

\item{word2}{second string}
}
\value{
The longest common substring from the end to \code{word1} and \code{word2} or \code{""} if there is no.
}
\description{
Find the longest common substring from the end between 2 words
}
\examples{
word1 <- "beautiful"
word2 <- "useful"
commonEnd(word1, word2)
}
