% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBasicFunctions.R
\name{commonStart}
\alias{commonStart}
\title{Find the longest common substring from the start between 2 words}
\usage{
commonStart(word1, word2)
}
\arguments{
\item{word1}{first string}

\item{word2}{second string}
}
\value{
The longest common substring from the start to \code{word1} and \code{word2} or \code{""} if there is no.
}
\description{
Find the longest common substring from the start between 2 words
}
\examples{
word1 <- "useless"
word2 <- "useful"
commonStart(word1, word2)
}
