% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBasicFunctions.R
\name{.readFileFromConditionOnNcols}
\alias{.readFileFromConditionOnNcols}
\title{Put the content of a tab separated file (with or without header gzip or not) in a dataframe
From the first line where the number of fields follow the condition cond}
\usage{
.readFileFromConditionOnNcols(fn, cond)
}
\arguments{
\item{fn}{the name of the file (tab delimited file with optionnally headers). Each row of the table appears as one line of the file. If it does not contain an absolute path, the file name is relative to the current working directory, getwd().}

\item{cond}{the condition that the number of columns should follow (for example "==4" or ">=3")}
}
\value{
The dataframe containing the values of the file \code{fn} (the header is removed).
}
\description{
Put the content of a tab separated file (with or without header gzip or not) in a dataframe
From the first line where the number of fields follow the condition cond
}
