% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGRBasicFunctions.R
\name{grFromBedFileWithName}
\alias{grFromBedFileWithName}
\title{create a GenomicRanges object from a BED}
\usage{
grFromBedFileWithName(fn)
}
\arguments{
\item{fn}{a path to the BED file (it can be gzipped) (0-based half open)}
}
\value{
a Genomic ranges with intervals (1-based closed)
}
\description{
create a GenomicRanges object from a BED
}
