% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBasicFunctions.R
\name{isValidColor}
\alias{isValidColor}
\title{Test if a variable could be used as a color}
\usage{
isValidColor(colorname)
}
\arguments{
\item{colorname}{a string or a numerical value containing the potential color}
}
\value{
A boolean which would say if the colorname is a valid color name or not.
}
\description{
Test if a variable could be used as a color
}
\details{
A valid color name is a numerical value or a string which is in colors() or a character vector with elements of 7 or 9 characters, "#" followed by the red, blue, green and optionally alpha values in hexadecimal (after rescaling to 0 ... 255).
}
