% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBasicFunctions.R
\name{rversionAbove}
\alias{rversionAbove}
\title{Test if the R version used is above the given version}
\usage{
rversionAbove(majorT, minorT = 0)
}
\arguments{
\item{majorT}{a string or numerical value of the major version to use as comparison (for example 3)}

\item{minorT}{a string or numerical value of the minor version to use as comparison (for example 5 or "5.0"). By default it is 0.}
}
\value{
\code{TRUE} if you are using a R version which is at least the one privided and \code{FALSE} if your version is below.
}
\description{
Test if the R version used is above the given version
}
