% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBasicFunctions.R
\name{simplifiedNamesByEnd}
\alias{simplifiedNamesByEnd}
\title{Simplify a vector of names removing the common end}
\usage{
simplifiedNamesByEnd(vecOfNames)
}
\arguments{
\item{vecOfNames}{a vector with different names to simplify}
}
\value{
a vector with the same length as \code{vecOfNames} where the end of the names were removed if they were identicals.
}
\description{
Simplify a vector of names removing the common end
}
\examples{
vecOfNames <- c("beautiful", "useful", "painful")
simplifiedNamesByEnd(vecOfNames)
}
