% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBasicFunctions.R
\name{simplifiedNamesByStart}
\alias{simplifiedNamesByStart}
\title{Simplify a vector of names removing the common end}
\usage{
simplifiedNamesByStart(vecOfNames)
}
\arguments{
\item{vecOfNames}{a vector with different names to simplify}
}
\value{
a vector with the same length as \code{vecOfNames} where the end of the names were removed if they were identicals.
}
\description{
Simplify a vector of names removing the common end
}
\examples{
vecOfNames <- c("old", "oldish", "older")
simplifiedNamesByStart(vecOfNames)
}
