% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBasicFunctions.R
\name{simplifiedNamesByStartEnd}
\alias{simplifiedNamesByStartEnd}
\title{Simplify a vector of names removing the common start and the common end}
\usage{
simplifiedNamesByStartEnd(vecOfNames)
}
\arguments{
\item{vecOfNames}{a vector with different names to simplify}
}
\value{
a vector with the same length as \code{vecOfNames} where the start and the end of the names were removed if they were identicals.
}
\description{
Simplify a vector of names removing the common start and the common end
}
\examples{
vecOfNames <- c("sampleOne.bed", "sampleTwo.bed", "sampleThree.bed")
simplifiedNamesByStartEnd(vecOfNames)
}
