% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBasicFunctions.R
\name{subsetByNamesOrIndices}
\alias{subsetByNamesOrIndices}
\title{Extract multiple elements from a list}
\usage{
subsetByNamesOrIndices(bigList, templNames)
}
\arguments{
\item{bigList}{a list from where elements will be extracted}

\item{templNames}{a vector containing indices or names of the element to extract}
}
\value{
A list containing only the elements named in \code{templNames} or whose indices are in \code{templNames} from \code{bigList}
}
\description{
Extract multiple elements from a list
}
\details{
The name of the output list will match the name of the \code{bigList} even if the indices were used.
}
\examples{
l<-list('a'=c(1,3),'b'=letters[1:10],'c'="fun")
subsetByNamesOrIndices(l,1:2)
subsetByNamesOrIndices(l,c("a", "c"))
}
