% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBasicFunctions.R
\name{cornerMat}
\alias{cornerMat}
\title{Show the corner of a matrix}
\usage{
cornerMat(matrix, size = 5, corner = "topleft", method = "minmax")
}
\arguments{
\item{matrix}{a matrix or a dataframe}

\item{size}{the size to display (default is 5)}

\item{corner}{the corner to display ("topleft", "topright", "bottomleft", or "bottomright"), default is "topleft"}

\item{method}{the method to select the top or bottom, default is 'minmax', can also be 'headtail'.}
}
\value{
a matrix or a dataframe `size` by `size`.
}
\description{
Show the corner of a matrix
}
\examples{
myHugeMatrix <- matrix(1:10000, nrow = 100)
cornerMat(myHugeMatrix, 10)
cornerMat(myHugeMatrix, 3, "bottomleft")
mySmallMatrix <- matrix(1:9, nrow = 3)
cornerMat(mySmallMatrix)
}
